client is now a useful thing
it is a management client and will interface with the server to control other clients. right now, it can only show the clients in a room. next, it will be able to display the actual state of each client. then, it will be able to filter clients by status and other such descriptors.