management clients communicate using TCP sockets
using TCP sockets are easier since they require you to stay connected to your peer. because of this i don't have to do anything to keep track of which clients are connected, and differetiate them from general clients. it makes it a lot easier to split up connections between each type of client.